Update limine version
The specific limine version used before had a bug where the trampoline
memory of the smp processors was freed too early and was therefore
marked as usable by the memory map. This failed when trying to write to
that memory space, as the cpu cores jumped to random places in memory.